+++ apple-codesign/src/code_resources.rs
@@ -1104,7 +1104,7 @@ impl CodeResourcesBuilder {
         &mut self,
         root_bundle_path: &Path,
         bundle_root: &Path,
-        context: &BundleSigningContext,
+        context: &mut BundleSigningContext,
     ) -> Result<(), AppleCodesignError> {
         let mut skipping_rel_dirs = BTreeSet::new();
 
@@ -1317,7 +1317,7 @@ impl CodeResourcesBuilder {
         rel_path: &Path,
         rel_path_normalized: &str,
         root_rel_path: &str,
-        context: &BundleSigningContext,
+        context: &mut BundleSigningContext,
         optional: bool,
     ) -> Result<(), AppleCodesignError> {
         let macho_info = if context
@@ -1352,12 +1352,11 @@ impl CodeResourcesBuilder {
         root_rel_path: &str,
         omit: bool,
         optional: bool,
-        context: &BundleSigningContext,
+        context: &mut BundleSigningContext,
     ) -> Result<(), AppleCodesignError> {
-        let mut need_install = true;
+        let mut need_install = !context.previously_installed_paths.contains(rel_path);
 
-        // Only seal if the omit flag is unset. But install unconditionally
-        // in all cases.
+        // Only seal if the omit flag is unset.
         if !omit {
             // Unlike Apple's tooling, we recognize Mach-O binaries when the nested
             // flag isn't set and we automatically sign.
@@ -1368,7 +1367,8 @@ impl CodeResourcesBuilder {
             // The reason we exclude in shallow mode is that shallow mode is supposed
             // to behave like Apple's `codesign` and that tool only signs the bundle's
             // "main" Mach-O binary, not other binaries.
-            let sign_macho = crate::reader::path_is_macho(full_path)?
+            let sign_macho = need_install
+                && crate::reader::path_is_macho(full_path)?
                 && !context
                     .settings
                     .path_exclusion_pattern_matches(root_rel_path)
@@ -1385,7 +1385,18 @@ impl CodeResourcesBuilder {
                 context.sign_and_install_macho(full_path, rel_path)?.0
             } else {
                 info!("sealing regular file {}", rel_path_normalized);
-                full_path.to_path_buf()
+
+                // If we need to install the file, seal the source file. Else since the
+                // file is already installed, seal the destination file.
+                //
+                // For regular files this distinction doesn't matter. But for Mach-O
+                // binaries it ensures we pick up the final signature, not the source
+                // file.
+                if need_install {
+                    full_path.to_path_buf()
+                } else {
+                    context.dest_dir.join(rel_path)
+                }
             };
 
             let digests = MultiDigest::from_path(read_path)?;
@@ -1415,7 +1426,7 @@ impl CodeResourcesBuilder {
         rel_path: &Path,
         rel_path_normalized: &str,
         omit: bool,
-        context: &BundleSigningContext,
+        context: &mut BundleSigningContext,
     ) -> Result<(), AppleCodesignError> {
         let link_target = std::fs::read_link(full_path)?
             .to_string_lossy()